refactor: make ParameterChangeForm a section. reuse DepositSection
