+++ src/components/ProposalForm.tsx
@@ -9,12 +9,16 @@ import { CodeInputGroup } from "./CodeInputGroup";
 import { CoreEval } from "@agoric/cosmic-proto/swingset/swingset.js";
 import { Button } from "./Button";
 import { ParamChange } from "cosmjs-types/cosmos/params/v1beta1/params";
+import { ParameterChangeFormSection } from "./ParameterChangeForm";
+import { DepositSection } from "./DepositSection";
 
-export type ProposalArgs = {
+type BaseProposalArgs = {
   title: string;
   description: string;
   deposit: string | number;
-} & ProposalDetail;
+};
+
+export type ProposalArgs = BaseProposalArgs & ProposalDetail;
 
 export type ProposalDetail =
   | { msgType: "textProposal" }
@@ -26,24 +30,28 @@ interface ProposalFormProps {
   description: string;
   handleSubmit: (proposal: ProposalArgs) => void;
   titleDescOnly?: boolean;
+  msgType: QueryParams["msgType"];
 }
 
 interface ProposalFormMethods {
   reset: () => void;
 }
 
-const msgType = "coreEvalProposal";
-
 const ProposalForm = forwardRef<ProposalFormMethods, ProposalFormProps>(
-  ({ title, description, handleSubmit, titleDescOnly = false }, ref) => {
+  ({ title, description, handleSubmit, msgType }, ref) => {
     const [evals, setEvals] = useState<CoreEval[]>([]);
     const formRef = useRef<HTMLFormElement>(null);
     const codeInputRef = useRef<{ reset: () => void } | null>(null);
+    const paramChangeRef = useRef<{
+      getChanges: () => ParamChange[];
+      reset: () => void;
+    } | null>(null);
 
     useImperativeHandle(ref, () => ({
       reset: () => {
         formRef.current?.reset();
         codeInputRef.current?.reset();
+        paramChangeRef.current?.reset();
         setEvals([]);
       },
     }));
@@ -55,8 +63,19 @@ const ProposalForm = forwardRef<ProposalFormMethods, ProposalFormProps>(
         if (formData) {
           const title = (formData.get("title") as string) || "";
           const description = (formData.get("description") as string) || "";
-          const deposit = (formData.get("deposit") as string) || "";
-          return handleSubmit({ title, description, deposit, msgType, evals });
+          const depositBld = (formData.get("deposit") as string) || "";
+          const deposit = Number(depositBld) * 1_000_000;
+          console.log("deposit", deposit);
+          const args: BaseProposalArgs = { title, description, deposit };
+          if (msgType === "coreEvalProposal" && evals.length) {
+            return handleSubmit({ ...args, msgType, evals });
+          } else if (msgType == "textProposal") {
+            return handleSubmit({ ...args, msgType });
+          } else if (msgType === "parameterChangeProposal") {
+            const changes = paramChangeRef.current?.getChanges();
+            if (!Array.isArray(changes)) throw new Error("No changes");
+            return handleSubmit({ ...args, msgType, changes });
+          }
         }
       }
       throw new Error("Error reading form data.");
@@ -115,7 +134,7 @@ const ProposalForm = forwardRef<ProposalFormMethods, ProposalFormProps>(
                 </div>
               </div>
 
-              {!titleDescOnly ? (
+              {msgType === "coreEvalProposal" ? (
                 <div className="sm:grid sm:grid-cols-4 sm:items-start sm:gap-4 sm:py-6">
                   <label
                     htmlFor="description"
@@ -132,31 +151,11 @@ const ProposalForm = forwardRef<ProposalFormMethods, ProposalFormProps>(
                 </div>
               ) : null}
 
-              <div className="sm:grid sm:grid-cols-4 sm:items-start sm:gap-4 sm:py-6">
-                <label
-                  htmlFor="description"
-                  className="block text-sm font-medium leading-6 text-gray-900 sm:pt-1.5"
-                >
-                  Deposit
-                </label>
-                <div className="mt-2 sm:col-span-3 sm:mt-0">
-                  <input
-                    type="number"
-                    min="0"
-                    step="1" // ensures integer
-                    defaultValue="0"
-                    name="deposit"
-                    id="deposit"
-                    className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-cardinal-600 sm:max-w-sm sm:text-sm sm:leading-6"
-                  />
-                  <p className="mt-3 text-sm leading-6 text-gray-600">
-                    A proposal requires{" "}
-                    <span className="font-semibold">10,000 ubld</span> to enter
-                    voting period. Current balance{" "}
-                    <span className="font-semibold">0 ubld</span>.
-                  </p>
-                </div>
-              </div>
+              {msgType === "parameterChangeProposal" ? (
+                <ParameterChangeFormSection ref={paramChangeRef} />
+              ) : null}
+
+              <DepositSection />
             </div>
           </div>
         </div>
